Integration tests from one folder
