--- integration-tests/asset/freeze_test.go
@@ -1,323 +1,0 @@
-//go:build integrationtests
-
-package asset
-
-import (
-	"testing"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	"github.com/CoreumFoundation/coreum/integration-tests"
-	"github.com/CoreumFoundation/coreum/pkg/tx"
-	"github.com/CoreumFoundation/coreum/testutil/event"
-	assettypes "github.com/CoreumFoundation/coreum/x/asset/types"
-)
-
-// TestFreezeUnfreezableFungibleToken checks freeze functionality on unfreezable fungible tokens.
-//
-//nolint:dupl
-func TestFreezeUnfreezableFungibleToken(t *testing.T) {
-	t.Parallel()
-
-	ctx, chain := integrationtests.NewTestingContext(t)
-
-	requireT := require.New(t)
-	assertT := assert.New(t)
-	issuer := chain.GenAccount()
-	recipient := chain.GenAccount()
-	requireT.NoError(
-		chain.Faucet.FundAccountsWithOptions(ctx, issuer, integrationtests.BalancesOptions{
-			Messages: []sdk.Msg{
-				&assettypes.MsgIssueFungibleToken{},
-				&assettypes.MsgFreezeFungibleToken{},
-			},
-		}))
-
-	// Issue an unfreezable fungible token
-	msg := &assettypes.MsgIssueFungibleToken{
-		Issuer:        issuer.String(),
-		Symbol:        "ABCNotFreezable",
-		Subunit:       "uabcnotfreezable",
-		Description:   "ABC Description",
-		Recipient:     recipient.String(),
-		InitialAmount: sdk.NewInt(1000),
-		Features:      []assettypes.FungibleTokenFeature{},
-	}
-
-	res, err := tx.BroadcastTx(
-		ctx,
-		chain.ClientContext.WithFromAddress(issuer),
-		chain.TxFactory().WithGas(chain.GasLimitByMsgs(msg)),
-		msg,
-	)
-
-	requireT.NoError(err)
-	fungibleTokenIssuedEvts, err := event.FindTypedEvents[*assettypes.EventFungibleTokenIssued](res.Events)
-	requireT.NoError(err)
-	unfreezableDenom := fungibleTokenIssuedEvts[0].Denom
-
-	// try to freeze unfreezable token
-	freezeMsg := &assettypes.MsgFreezeFungibleToken{
-		Sender:  issuer.String(),
-		Account: recipient.String(),
-		Coin:    sdk.NewCoin(unfreezableDenom, sdk.NewInt(1000)),
-	}
-	_, err = tx.BroadcastTx(
-		ctx,
-		chain.ClientContext.WithFromAddress(issuer),
-		chain.TxFactory().WithGas(chain.GasLimitByMsgs(freezeMsg)),
-		freezeMsg,
-	)
-	assertT.True(assettypes.ErrFeatureNotActive.Is(err))
-}
-
-// TestFreezeFungibleToken checks freeze functionality of fungible tokens.
-func TestFreezeFungibleToken(t *testing.T) {
-	t.Parallel()
-
-	ctx, chain := integrationtests.NewTestingContext(t)
-
-	requireT := require.New(t)
-	assertT := assert.New(t)
-	clientCtx := chain.ClientContext
-
-	assetClient := assettypes.NewQueryClient(clientCtx)
-	bankClient := banktypes.NewQueryClient(clientCtx)
-
-	issuer := chain.GenAccount()
-	recipient := chain.GenAccount()
-	randomAddress := chain.GenAccount()
-	requireT.NoError(
-		chain.Faucet.FundAccountsWithOptions(ctx, issuer, integrationtests.BalancesOptions{
-			Messages: []sdk.Msg{
-				&assettypes.MsgIssueFungibleToken{},
-				&assettypes.MsgIssueFungibleToken{},
-				&assettypes.MsgFreezeFungibleToken{},
-				&assettypes.MsgFreezeFungibleToken{},
-				&assettypes.MsgUnfreezeFungibleToken{},
-				&assettypes.MsgUnfreezeFungibleToken{},
-				&assettypes.MsgUnfreezeFungibleToken{},
-			},
-		}))
-	requireT.NoError(
-		chain.Faucet.FundAccountsWithOptions(ctx, recipient, integrationtests.BalancesOptions{
-			Messages: []sdk.Msg{
-				&banktypes.MsgSend{},
-				&banktypes.MsgSend{},
-				&banktypes.MsgSend{},
-				&banktypes.MsgSend{},
-			},
-		}))
-	requireT.NoError(
-		chain.Faucet.FundAccountsWithOptions(ctx, randomAddress, integrationtests.BalancesOptions{
-			Messages: []sdk.Msg{
-				&assettypes.MsgFreezeFungibleToken{},
-			},
-		}))
-
-	// Issue the new fungible token
-	msg := &assettypes.MsgIssueFungibleToken{
-		Issuer:        issuer.String(),
-		Symbol:        "ABC",
-		Subunit:       "uabc",
-		Precision:     6,
-		Description:   "ABC Description",
-		Recipient:     recipient.String(),
-		InitialAmount: sdk.NewInt(1000),
-		Features: []assettypes.FungibleTokenFeature{
-			assettypes.FungibleTokenFeature_freeze, //nolint:nosnakecase
-		},
-	}
-
-	res, err := tx.BroadcastTx(
-		ctx,
-		chain.ClientContext.WithFromAddress(issuer),
-		chain.TxFactory().WithGas(chain.GasLimitByMsgs(msg)),
-		msg,
-	)
-
-	requireT.NoError(err)
-	fungibleTokenIssuedEvts, err := event.FindTypedEvents[*assettypes.EventFungibleTokenIssued](res.Events)
-	requireT.NoError(err)
-	denom := fungibleTokenIssuedEvts[0].Denom
-
-	// try to pass non-issuer signature to freeze msg
-	freezeMsg := &assettypes.MsgFreezeFungibleToken{
-		Sender:  randomAddress.String(),
-		Account: recipient.String(),
-		Coin:    sdk.NewCoin(denom, sdk.NewInt(1000)),
-	}
-	_, err = tx.BroadcastTx(
-		ctx,
-		chain.ClientContext.WithFromAddress(randomAddress),
-		chain.TxFactory().WithGas(chain.GasLimitByMsgs(freezeMsg)),
-		freezeMsg,
-	)
-	requireT.Error(err)
-	assertT.True(sdkerrors.ErrUnauthorized.Is(err))
-
-	// freeze 400 tokens
-	freezeMsg = &assettypes.MsgFreezeFungibleToken{
-		Sender:  issuer.String(),
-		Account: recipient.String(),
-		Coin:    sdk.NewCoin(denom, sdk.NewInt(400)),
-	}
-	res, err = tx.BroadcastTx(
-		ctx,
-		chain.ClientContext.WithFromAddress(issuer),
-		chain.TxFactory().WithGas(chain.GasLimitByMsgs(freezeMsg)),
-		freezeMsg,
-	)
-	requireT.NoError(err)
-	assertT.EqualValues(res.GasUsed, chain.GasLimitByMsgs(freezeMsg))
-
-	fungibleTokenFreezeEvts, err := event.FindTypedEvents[*assettypes.EventFungibleTokenFrozenAmountChanged](res.Events)
-	requireT.NoError(err)
-	assertT.EqualValues(&assettypes.EventFungibleTokenFrozenAmountChanged{
-		Account:        recipient.String(),
-		PreviousAmount: sdk.NewCoin(denom, sdk.NewInt(0)),
-		CurrentAmount:  sdk.NewCoin(denom, sdk.NewInt(400)),
-	}, fungibleTokenFreezeEvts[0])
-
-	// query frozen tokens
-	frozenBalance, err := assetClient.FrozenBalance(ctx, &assettypes.QueryFrozenBalanceRequest{
-		Account: recipient.String(),
-		Denom:   denom,
-	})
-	requireT.NoError(err)
-	requireT.EqualValues(sdk.NewCoin(denom, sdk.NewInt(400)), frozenBalance.Balance)
-
-	frozenBalances, err := assetClient.FrozenBalances(ctx, &assettypes.QueryFrozenBalancesRequest{
-		Account: recipient.String(),
-	})
-	requireT.NoError(err)
-	requireT.EqualValues(sdk.NewCoins(sdk.NewCoin(denom, sdk.NewInt(400))), frozenBalances.Balances)
-
-	// try to send more than available (650) (600 is available)
-	recipient2 := chain.GenAccount()
-	sendMsg := &banktypes.MsgSend{
-		FromAddress: recipient.String(),
-		ToAddress:   recipient2.String(),
-		Amount:      sdk.NewCoins(sdk.NewCoin(denom, sdk.NewInt(650))),
-	}
-	_, err = tx.BroadcastTx(
-		ctx,
-		chain.ClientContext.WithFromAddress(recipient),
-		chain.TxFactory().WithGas(chain.GasLimitByMsgs(sendMsg)),
-		sendMsg,
-	)
-	requireT.Error(err)
-	assertT.True(sdkerrors.ErrInsufficientFunds.Is(err))
-
-	// try to send available tokens (600)
-	sendMsg = &banktypes.MsgSend{
-		FromAddress: recipient.String(),
-		ToAddress:   recipient2.String(),
-		Amount:      sdk.NewCoins(sdk.NewCoin(denom, sdk.NewInt(600))),
-	}
-	_, err = tx.BroadcastTx(
-		ctx,
-		chain.ClientContext.WithFromAddress(recipient),
-		chain.TxFactory().WithGas(chain.GasLimitByMsgs(sendMsg)),
-		sendMsg,
-	)
-	requireT.NoError(err)
-	balance1, err := bankClient.Balance(ctx, &banktypes.QueryBalanceRequest{
-		Address: recipient.String(),
-		Denom:   denom,
-	})
-	requireT.NoError(err)
-	requireT.Equal(sdk.NewCoin(denom, sdk.NewInt(400)).String(), balance1.GetBalance().String())
-
-	// unfreeze 200 tokens and try send 250 tokens
-	unfreezeMsg := &assettypes.MsgUnfreezeFungibleToken{
-		Sender:  issuer.String(),
-		Account: recipient.String(),
-		Coin:    sdk.NewCoin(denom, sdk.NewInt(200)),
-	}
-	res, err = tx.BroadcastTx(
-		ctx,
-		chain.ClientContext.WithFromAddress(issuer),
-		chain.TxFactory().WithGas(chain.GasLimitByMsgs(unfreezeMsg)),
-		unfreezeMsg,
-	)
-	requireT.NoError(err)
-	assertT.EqualValues(res.GasUsed, chain.GasLimitByMsgs(unfreezeMsg))
-
-	fungibleTokenFreezeEvts, err = event.FindTypedEvents[*assettypes.EventFungibleTokenFrozenAmountChanged](res.Events)
-	requireT.NoError(err)
-	assertT.EqualValues(&assettypes.EventFungibleTokenFrozenAmountChanged{
-		Account:        recipient.String(),
-		PreviousAmount: sdk.NewCoin(denom, sdk.NewInt(400)),
-		CurrentAmount:  sdk.NewCoin(denom, sdk.NewInt(200)),
-	}, fungibleTokenFreezeEvts[0])
-
-	sendMsg = &banktypes.MsgSend{
-		FromAddress: recipient.String(),
-		ToAddress:   recipient2.String(),
-		Amount:      sdk.NewCoins(sdk.NewCoin(denom, sdk.NewInt(250))),
-	}
-	_, err = tx.BroadcastTx(
-		ctx,
-		chain.ClientContext.WithFromAddress(recipient),
-		chain.TxFactory().WithGas(chain.GasLimitByMsgs(sendMsg)),
-		sendMsg,
-	)
-	requireT.Error(err)
-	assertT.True(sdkerrors.ErrInsufficientFunds.Is(err))
-
-	// send available tokens (200)
-	sendMsg = &banktypes.MsgSend{
-		FromAddress: recipient.String(),
-		ToAddress:   recipient2.String(),
-		Amount:      sdk.NewCoins(sdk.NewCoin(denom, sdk.NewInt(200))),
-	}
-	_, err = tx.BroadcastTx(
-		ctx,
-		chain.ClientContext.WithFromAddress(recipient),
-		chain.TxFactory().WithGas(chain.GasLimitByMsgs(sendMsg)),
-		sendMsg,
-	)
-	requireT.NoError(err)
-
-	// unfreeze 400 tokens (frozen balance is 200), it should give error
-	unfreezeMsg = &assettypes.MsgUnfreezeFungibleToken{
-		Sender:  issuer.String(),
-		Account: recipient.String(),
-		Coin:    sdk.NewCoin(denom, sdk.NewInt(400)),
-	}
-	_, err = tx.BroadcastTx(
-		ctx,
-		chain.ClientContext.WithFromAddress(issuer),
-		chain.TxFactory().WithGas(chain.GasLimitByMsgs(unfreezeMsg)),
-		unfreezeMsg,
-	)
-	requireT.True(assettypes.ErrNotEnoughBalance.Is(err))
-
-	// unfreeze 200 tokens and observer current frozen amount is zero
-	unfreezeMsg = &assettypes.MsgUnfreezeFungibleToken{
-		Sender:  issuer.String(),
-		Account: recipient.String(),
-		Coin:    sdk.NewCoin(denom, sdk.NewInt(200)),
-	}
-	res, err = tx.BroadcastTx(
-		ctx,
-		chain.ClientContext.WithFromAddress(issuer),
-		chain.TxFactory().WithGas(chain.GasLimitByMsgs(unfreezeMsg)),
-		unfreezeMsg,
-	)
-	requireT.NoError(err)
-	assertT.EqualValues(res.GasUsed, chain.GasLimitByMsgs(unfreezeMsg))
-
-	fungibleTokenFreezeEvts, err = event.FindTypedEvents[*assettypes.EventFungibleTokenFrozenAmountChanged](res.Events)
-	requireT.NoError(err)
-	assertT.EqualValues(&assettypes.EventFungibleTokenFrozenAmountChanged{
-		Account:        recipient.String(),
-		PreviousAmount: sdk.NewCoin(denom, sdk.NewInt(200)),
-		CurrentAmount:  sdk.NewCoin(denom, sdk.NewInt(0)),
-	}, fungibleTokenFreezeEvts[0])
-}
