--- integration-tests/auth/sequence_test.go
@@ -1,49 +1,0 @@
-//go:build integrationtests
-
-package auth
-
-import (
-	"testing"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	cosmoserrors "github.com/cosmos/cosmos-sdk/types/errors"
-	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
-	"github.com/stretchr/testify/require"
-
-	"github.com/CoreumFoundation/coreum/integration-tests"
-	"github.com/CoreumFoundation/coreum/pkg/tx"
-)
-
-// TestUnexpectedSequenceNumber test verifies that we correctly handle error reporting invalid account sequence number
-// used to sign transaction
-func TestUnexpectedSequenceNumber(t *testing.T) {
-	t.Parallel()
-
-	ctx, chain := integrationtests.NewTestingContext(t)
-
-	sender := chain.GenAccount()
-
-	require.NoError(t, chain.Faucet.FundAccountsWithOptions(ctx, sender, integrationtests.BalancesOptions{
-		Messages: []sdk.Msg{&banktypes.MsgSend{}},
-		Amount:   sdk.NewInt(10),
-	}))
-
-	clientCtx := chain.ClientContext
-	accInfo, err := tx.GetAccountInfo(ctx, clientCtx, sender)
-	require.NoError(t, err)
-
-	msg := &banktypes.MsgSend{
-		FromAddress: sender.String(),
-		ToAddress:   sender.String(),
-		Amount:      sdk.NewCoins(chain.NewCoin(sdk.NewInt(1))),
-	}
-
-	_, err = tx.BroadcastTx(ctx,
-		chain.ClientContext.WithFromAddress(sender),
-		chain.TxFactory().
-			WithSequence(accInfo.GetSequence()+1). // incorrect sequence
-			WithAccountNumber(accInfo.GetAccountNumber()).
-			WithGas(chain.GasLimitByMsgs(msg)),
-		msg)
-	require.True(t, cosmoserrors.ErrWrongSequence.Is(err))
-}
