+++ integration-tests/modules/auth_test.go
@@ -1,6 +1,6 @@
 //go:build integrationtests
 
-package auth
+package modules
 
 import (
 	"testing"
@@ -10,6 +10,7 @@ import (
 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	multisigtypes "github.com/cosmos/cosmos-sdk/crypto/types/multisig"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	cosmoserrors "github.com/cosmos/cosmos-sdk/types/errors"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	sdksigning "github.com/cosmos/cosmos-sdk/types/tx/signing"
 	authsigning "github.com/cosmos/cosmos-sdk/x/auth/signing"
@@ -18,9 +19,70 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/CoreumFoundation/coreum-tools/pkg/logger"
-	"github.com/CoreumFoundation/coreum/integration-tests"
+	integrationtests "github.com/CoreumFoundation/coreum/integration-tests"
 	"github.com/CoreumFoundation/coreum/pkg/tx"
 )
+
+// TODO (wojtek): once we have other coins add test verifying that transaction offering fee in coin other then CORE is rejected
+
+// TestFeeLimits verifies that invalid message gas won't be accepted.
+func TestFeeLimits(t *testing.T) {
+	t.Parallel()
+
+	ctx, chain := integrationtests.NewTestingContext(t)
+
+	sender := chain.GenAccount()
+
+	maxBlockGas := chain.NetworkConfig.Fee.FeeModel.Params().MaxBlockGas
+	require.NoError(t, chain.Faucet.FundAccountsWithOptions(ctx, sender, integrationtests.BalancesOptions{
+		Messages: []sdk.Msg{&banktypes.MsgSend{}},
+		Amount:   sdk.NewInt(maxBlockGas + 100),
+	}))
+
+	msg := &banktypes.MsgSend{
+		FromAddress: sender.String(),
+		ToAddress:   sender.String(),
+		Amount:      sdk.NewCoins(chain.NewCoin(sdk.NewInt(1))),
+	}
+
+	gasPriceWithMaxDiscount := chain.NetworkConfig.Fee.FeeModel.Params().InitialGasPrice.
+		Mul(sdk.OneDec().Sub(chain.NetworkConfig.Fee.FeeModel.Params().MaxDiscount))
+
+	// the gas price is too low
+	_, err := tx.BroadcastTx(ctx,
+		chain.ClientContext.WithFromAddress(sender),
+		chain.TxFactory().
+			WithGas(chain.GasLimitByMsgs(msg)).
+			WithGasPrices(chain.NewDecCoin(gasPriceWithMaxDiscount.QuoInt64(2)).String()),
+		msg)
+	require.True(t, cosmoserrors.ErrInsufficientFee.Is(err))
+
+	// no gas price
+	_, err = tx.BroadcastTx(ctx,
+		chain.ClientContext.WithFromAddress(sender),
+		chain.TxFactory().
+			WithGas(chain.GasLimitByMsgs(msg)).
+			WithGasPrices(""),
+		msg)
+	require.True(t, cosmoserrors.ErrInsufficientFee.Is(err))
+
+	// more gas than MaxBlockGas
+	_, err = tx.BroadcastTx(ctx,
+		chain.ClientContext.WithFromAddress(sender),
+		chain.TxFactory().
+			WithGas(uint64(maxBlockGas+1)),
+		msg)
+	// TODO(dhil) here we get the Internal error -> "tx (***) not found" and the test takes the "txTimeout" time, validate that it's expected
+	require.Error(t, err)
+
+	// gas equal MaxBlockGas, the tx should pass
+	_, err = tx.BroadcastTx(ctx,
+		chain.ClientContext.WithFromAddress(sender),
+		chain.TxFactory().
+			WithGas(uint64(maxBlockGas)),
+		msg)
+	require.NoError(t, err)
+}
 
 // TestMultisig tests the cosmos-sdk multisig accounts and API.
 func TestMultisig(t *testing.T) {
@@ -117,6 +179,40 @@ func TestMultisig(t *testing.T) {
 	})
 	requireT.NoError(err)
 	requireT.Equal(coinsToSendToRecipient, recipientBalances.Balances)
+}
+
+// TestUnexpectedSequenceNumber test verifies that we correctly handle error reporting invalid account sequence number
+// used to sign transaction
+func TestUnexpectedSequenceNumber(t *testing.T) {
+	t.Parallel()
+
+	ctx, chain := integrationtests.NewTestingContext(t)
+
+	sender := chain.GenAccount()
+
+	require.NoError(t, chain.Faucet.FundAccountsWithOptions(ctx, sender, integrationtests.BalancesOptions{
+		Messages: []sdk.Msg{&banktypes.MsgSend{}},
+		Amount:   sdk.NewInt(10),
+	}))
+
+	clientCtx := chain.ClientContext
+	accInfo, err := tx.GetAccountInfo(ctx, clientCtx, sender)
+	require.NoError(t, err)
+
+	msg := &banktypes.MsgSend{
+		FromAddress: sender.String(),
+		ToAddress:   sender.String(),
+		Amount:      sdk.NewCoins(chain.NewCoin(sdk.NewInt(1))),
+	}
+
+	_, err = tx.BroadcastTx(ctx,
+		chain.ClientContext.WithFromAddress(sender),
+		chain.TxFactory().
+			WithSequence(accInfo.GetSequence()+1). // incorrect sequence
+			WithAccountNumber(accInfo.GetAccountNumber()).
+			WithGas(chain.GasLimitByMsgs(msg)),
+		msg)
+	require.True(t, cosmoserrors.ErrWrongSequence.Is(err))
 }
 
 func createMulisignTx(requireT *require.Assertions, txBuilder sdkclient.TxBuilder, accSec uint64, multisigPublicKey *sdkmultisig.LegacyAminoPubKey) authsigning.Tx {
