+++ integration-tests/modules/staking_test.go
@@ -1,10 +1,11 @@
 //go:build integrationtests
 
-package staking
+package modules
 
 import (
 	"context"
 	"fmt"
+	"strconv"
 	"testing"
 	"time"
 
@@ -19,7 +20,7 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/CoreumFoundation/coreum-tools/pkg/logger"
-	"github.com/CoreumFoundation/coreum/integration-tests"
+	integrationtests "github.com/CoreumFoundation/coreum/integration-tests"
 	"github.com/CoreumFoundation/coreum/pkg/tx"
 	customparamstypes "github.com/CoreumFoundation/coreum/x/customparams/types"
 )
@@ -265,6 +266,57 @@ func TestValidatorUpdateWithLowMinSelfDelegation(t *testing.T) {
 	assert.EqualValues(t, editValidatorMsg.Description.Details, valResp.GetValidator().Description.Details)
 }
 
+// TestStakingProposalParamChange checks that staking param change proposal works correctly.
+func TestStakingProposalParamChange(t *testing.T) {
+	t.Parallel()
+
+	ctx, chain := integrationtests.NewTestingContext(t)
+
+	targetMaxValidators := 2 * chain.NetworkConfig.StakingConfig.MaxValidators
+	requireT := require.New(t)
+	stakingClient := stakingtypes.NewQueryClient(chain.ClientContext)
+
+	// Create new proposer.
+	proposer := chain.GenAccount()
+	proposerBalance, err := chain.Governance.ComputeProposerBalance(ctx)
+	requireT.NoError(err)
+
+	err = chain.Faucet.FundAccounts(ctx, integrationtests.NewFundedAccount(proposer, proposerBalance))
+	requireT.NoError(err)
+
+	// Create proposition to change max validators value.
+	proposalMsg, err := chain.Governance.NewMsgSubmitProposal(ctx, proposer, paramproposal.NewParameterChangeProposal("Change MaxValidators", "Propose changing MaxValidators in the staking module",
+		[]paramproposal.ParamChange{
+			paramproposal.NewParamChange(stakingtypes.ModuleName, string(stakingtypes.KeyMaxValidators), strconv.Itoa(targetMaxValidators)),
+		},
+	))
+	requireT.NoError(err)
+	proposalID, err := chain.Governance.Propose(ctx, proposalMsg)
+	requireT.NoError(err)
+	logger.Get(ctx).Info("Proposal has been submitted", zap.Uint64("proposalID", proposalID))
+
+	// Verify that voting period started.
+	proposal, err := chain.Governance.GetProposal(ctx, proposalID)
+	requireT.NoError(err)
+	requireT.Equal(govtypes.StatusVotingPeriod, proposal.Status)
+
+	// Vote yes from all vote accounts.
+	err = chain.Governance.VoteAll(ctx, govtypes.OptionYes, proposal.ProposalId)
+	requireT.NoError(err)
+
+	logger.Get(ctx).Info("Voters have voted successfully, waiting for voting period to be finished", zap.Time("votingEndTime", proposal.VotingEndTime))
+
+	// Wait for proposal result.
+	finalStatus, err := chain.Governance.WaitForVotingToFinalize(ctx, proposalID)
+	requireT.NoError(err)
+	requireT.Equal(govtypes.StatusPassed, finalStatus)
+
+	// Check the proposed change is applied.
+	stakingParams, err := stakingClient.Params(ctx, &stakingtypes.QueryParamsRequest{})
+	requireT.NoError(err)
+	requireT.Equal(uint32(targetMaxValidators), stakingParams.Params.MaxValidators)
+}
+
 func changeMinSelfDelegationCustomParam(
 	ctx context.Context,
 	requireT *require.Assertions,
@@ -315,7 +367,7 @@ func setUnbondingTimeViaGovernance(ctx context.Context, t *testing.T, chain inte
 	err = chain.Faucet.FundAccounts(ctx, integrationtests.NewFundedAccount(proposer, proposerBalance))
 	requireT.NoError(err)
 
-	// TODO(dhil) refactor other integrationtests to use that func for the standard propose + vote action.
+	// TODO(dhil) refactor other tests to use that func for the standard propose + vote action.
 	// Create proposition to change max the unbonding time value.
 	err = chain.Governance.ProposeAndVote(ctx, proposer,
 		paramproposal.NewParameterChangeProposal(
