+++ integration-tests/modules/testdata/wasm/.gitignore
