+++ integration-tests/modules/testdata/wasm/bank-send/Cargo.toml
