+++ integration-tests/modules/testdata/wasm/bank-send/artifacts/bank_send.wasm
