+++ integration-tests/modules/testdata/wasm/bank-send/artifacts/checksums.txt
