+++ integration-tests/modules/testdata/wasm/bank-send/src/contract.rs
