+++ integration-tests/modules/testdata/wasm/bank-send/src/lib.rs
