+++ integration-tests/modules/testdata/wasm/issue-fungible-token/Cargo.toml
