+++ integration-tests/modules/testdata/wasm/issue-fungible-token/artifacts/checksums.txt
