+++ integration-tests/modules/testdata/wasm/issue-fungible-token/artifacts/issue_fungible_token.wasm
