+++ integration-tests/modules/testdata/wasm/issue-fungible-token/src/lib.rs
