+++ integration-tests/modules/testdata/wasm/issue-fungible-token/src/sdk.rs
