+++ integration-tests/modules/testdata/wasm/simple-state/Cargo.toml
