+++ integration-tests/modules/testdata/wasm/simple-state/artifacts/checksums.txt
