+++ integration-tests/modules/testdata/wasm/simple-state/artifacts/simple_state.wasm
