+++ integration-tests/modules/testdata/wasm/simple-state/src/contract.rs
