+++ integration-tests/modules/testdata/wasm/simple-state/src/lib.rs
