+++ integration-tests/staking.go
@@ -15,7 +15,7 @@ import (
 )
 
 // CreateValidator creates a new validator on the chain and returns the staker addresses, validator addresses and callback function to deactivate it.
-func CreateValidator(ctx context.Context, chain Chain, stakingAmount sdk.Int, selfDelegationAmount sdk.Int) (sdk.AccAddress, sdk.ValAddress, func() error, error) {
+func CreateValidator(ctx context.Context, chain Chain, stakingAmount, selfDelegationAmount sdk.Int) (sdk.AccAddress, sdk.ValAddress, func() error, error) {
 	stakingClient := stakingtypes.NewQueryClient(chain.ClientContext)
 	staker := chain.GenAccount()
