--- integration-tests/wasm/bank_send_test.go
@@ -1,148 +1,0 @@
-//go:build integrationtests
-
-package wasm
-
-import (
-	_ "embed"
-	"encoding/json"
-	"testing"
-
-	wasmtypes "github.com/CosmWasm/wasmd/x/wasm/types"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	cosmoserrors "github.com/cosmos/cosmos-sdk/types/errors"
-	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
-	"github.com/stretchr/testify/require"
-
-	"github.com/CoreumFoundation/coreum/integration-tests"
-	"github.com/CoreumFoundation/coreum/pkg/tx"
-)
-
-var (
-	//go:embed testdata/bank-send/artifacts/bank_send.wasm
-	bankSendWASM []byte
-)
-
-type bankWithdrawRequest struct {
-	Amount    string `json:"amount"`
-	Denom     string `json:"denom"`
-	Recipient string `json:"recipient"`
-}
-
-type bankMethod string
-
-const (
-	withdraw bankMethod = "withdraw"
-)
-
-// TestBankSendWASMContract runs a contract deployment flow and tests that the contract is able to use Bank module
-// to disperse the native coins.
-func TestBankSendWASMContract(t *testing.T) {
-	t.Parallel()
-
-	ctx, chain := integrationtests.NewTestingContext(t)
-
-	admin := chain.GenAccount()
-	nativeDenom := chain.NetworkConfig.Denom
-
-	requireT := require.New(t)
-	requireT.NoError(chain.Faucet.FundAccounts(ctx,
-		integrationtests.NewFundedAccount(admin, chain.NewCoin(sdk.NewInt(5000000000))),
-	))
-
-	clientCtx := chain.ClientContext.WithFromAddress(admin)
-	txf := chain.TxFactory().
-		WithSimulateAndExecute(true)
-	bankClient := banktypes.NewQueryClient(clientCtx)
-
-	// deploy and init contract with the initial coins amount
-	initialPayload, err := json.Marshal(struct{}{})
-	requireT.NoError(err)
-	contractAddr, _, err := DeployAndInstantiate(
-		ctx,
-		clientCtx,
-		txf,
-		bankSendWASM,
-		InstantiateConfig{
-			accessType: wasmtypes.AccessTypeUnspecified,
-			payload:    initialPayload,
-			amount:     chain.NewCoin(sdk.NewInt(10000)),
-			label:      "bank_send",
-		},
-	)
-	requireT.NoError(err)
-
-	// send additional coins to contract directly
-	sdkContractAddress, err := sdk.AccAddressFromBech32(contractAddr)
-	requireT.NoError(err)
-
-	msg := &banktypes.MsgSend{
-		FromAddress: admin.String(),
-		ToAddress:   sdkContractAddress.String(),
-		Amount:      sdk.NewCoins(chain.NewCoin(sdk.NewInt(5000))),
-	}
-
-	_, err = tx.BroadcastTx(ctx, clientCtx, txf, msg)
-	requireT.NoError(err)
-
-	// get the contract balance and check total
-	contractBalance, err := bankClient.Balance(ctx,
-		&banktypes.QueryBalanceRequest{
-			Address: contractAddr,
-			Denom:   nativeDenom,
-		})
-	requireT.NoError(err)
-	requireT.NotNil(contractBalance.Balance)
-	requireT.Equal(sdk.NewInt64Coin(nativeDenom, 15000).String(), contractBalance.Balance.String())
-
-	recipient := chain.GenAccount()
-	// try to exceed the contract limit
-	withdrawPayload, err := json.Marshal(map[bankMethod]bankWithdrawRequest{
-		withdraw: {
-			Amount:    "16000",
-			Denom:     nativeDenom,
-			Recipient: recipient.String(),
-		},
-	})
-	requireT.NoError(err)
-
-	// try to withdraw more than the admin has
-	txf = txf.
-		WithSimulateAndExecute(false).
-		// the gas here is to try to execute the tx and don't fail on the gas estimation
-		WithGas(uint64(chain.NetworkConfig.Fee.FeeModel.Params().MaxBlockGas))
-	_, err = Execute(ctx, clientCtx, txf, contractAddr, withdrawPayload, sdk.Coin{})
-	requireT.True(cosmoserrors.ErrInsufficientFunds.Is(err))
-
-	// send coin from the contract to test wallet
-	withdrawPayload, err = json.Marshal(map[bankMethod]bankWithdrawRequest{
-		withdraw: {
-			Amount:    "5000",
-			Denom:     nativeDenom,
-			Recipient: recipient.String(),
-		},
-	})
-	requireT.NoError(err)
-
-	txf = txf.WithSimulateAndExecute(true)
-	_, err = Execute(ctx, clientCtx, txf, contractAddr, withdrawPayload, sdk.Coin{})
-	requireT.NoError(err)
-
-	// check contract and wallet balances
-	contractBalance, err = bankClient.Balance(ctx,
-		&banktypes.QueryBalanceRequest{
-			Address: contractAddr,
-			Denom:   nativeDenom,
-		})
-	requireT.NoError(err)
-	requireT.NotNil(contractBalance.Balance)
-	requireT.Equal(sdk.NewInt64Coin(nativeDenom, 10000).String(), contractBalance.Balance.String())
-
-	recipientBalance, err := bankClient.Balance(ctx,
-		&banktypes.QueryBalanceRequest{
-			Address: recipient.String(),
-			Denom:   nativeDenom,
-		})
-	requireT.NoError(err)
-	requireT.NotNil(recipientBalance.Balance)
-	requireT.Equal(sdk.NewInt64Coin(nativeDenom, 5000).String(), recipientBalance.Balance.String())
-}
